Rename: Focus-Window -> Use-Window (extension ID accordingly)
"Focus" is not an approved Powershell verb.[1] The original Focus-Window
PowerShell module is using it anyway and recommends the addition of the
"-DisableNameChecking" parameter when importing the module.[2]

However, this is not an option in a Chocolatey context, especially as
this will trigger warnings on every "choco" call. I therefore renamed
the function to "Use-Window" and the whole Chocolatey extension
accordingly.

[1] https://learn.microsoft.com/en-us/powershell/scripting/developer/cmdlet/approved-verbs-for-windows-powershell-commands

[2] From https://github.com/71/Focus-Window#installation:

    > Note the addition of -DisableNameChecking, which is required since
    > Focus-Window begins with "Focus", which is not an approved
    > PowerShell verb.